--- components/README.txt
@@ -1,1 +1,0 @@
-This section will require separate Arduino microcontrollers
