+++ home-security-system.ino
@@ -1,122 +1,106 @@
-#include <Wire.h>
 
-int systemStatus = 0; // status to be reflected in the sequence
-int i2cID = 9; //slave address
-int buttonPin = 7;
-int signalInterval = 10;
-int signalIntervalDefault = 10;
+bool setAlarm = LOW; // global check for alarm
+bool systemUp = HIGH;
 
-/*
-ACTUATOR MODULES
- */
-#include "modules/buzzer.h"
-#include "modules/segmentLED.h"
-#include "modules/relaySwitch.h"
-#include "modules/ethernetRelay.h"
-////////////////////////////////////////// alarm status
-void setAlarmOn(int x){
-	Serial.println("Alarm|Alarm|Alarm|Alarm|Alarm|Alarm");
-	// soundBuzzerOn();
-  // relaySwitchOn();
-  sendEthernetSignal(x);
-}
-void setAlarmOff(){
-  Serial.println("Alarm Off");
-  soundBuzzerOff();
-  relaySwitchOff();
-  sendEthernetSignal(0);
-}
-void initAlarm(){
-  initCodeLights();
-  //initBuzzer();
-  initI2CSlave();
-  //initRelaySwitch();
-  initEthernetRelay();
-}
-void setStatus(int x){
-	int i = x % 16;
-	Serial.print("at status: ");
-	Serial.println(i);
-	setLEDStatus(i);
-	if(i == 15){
-		setAlarmOn(i);
-   delay(1000);
-	}
-}
+//------- PIN REGISTRY
+// 3 - vibratePin   : IN  : 
+// 4 - laserPin    : OUT : laser light
+// 5 - uhEchoPin    : IN  : Ultrasonic echo pin
+// 6 - uhTrigPin    : OUT : Ultrasonic trigger pin
+// 7 - dht11Pin     : IN  : Digital Humidity and Temperature
+// 9 - speakerPin   : OUT : Speaker alarm pin
+// 11 - resetButton : IN  : set Alarm to OFF state
+// 12 - greenLED    : OUT : green LED light
+// 13 - redLED      : OUT : red LED light
 
-void initI2CSlave(){
-  Wire.begin(i2cID); 
-  Wire.onReceive(statusReceiver);
-}
-void statusReceiver(int bytes){
-  if(systemStatus < 15){
-  	int x = Wire.read();    // read one character from the I2C
-  	Serial.print("........................reading wire : ");
-  	Serial.println(x);
-  	systemStatus = x;
-  	setStatus(x);
-  }
-  else{
-    Serial.println("system is on alarm");
-    setAlarmOn(15);
-  }
-	delay(100);
-}
-/**
-SENSOR MODULES: 
-	attached to the Arduino Mega
-	defined and configured on /modules/
-	requires an init<Module>() and check<Module>() 
-		to be included in the initModules() and checkModules() below
-*/
+// 25
+
+// 45 - openReedPin : IN  : Magnetic switch
+// 46 - codePins
+// 47 - codePins
+// 48 - codePins
+// 49 - codePins
+// 50 - codePins
+// 51 - codePins
+// 52 - codePins
+// 53 - codePins
+
+// A0 - ambientLightPin : ambient light sensor/ laser receiver
+// A1 - tempPin     : temperature sensor
+// A2 - flamePin    : Flam sensor
+
+//------ NUMBER CODE REGISTRY
+// 0
+// 1 - switch
+// 2 - dht Temp
+// 3 - flame
+// 4 - laser trip
+// 6 - vibrate
+// 7 - reed/magnet
+// 8 - temperature
+// 9 - untrasonic 
+
+
+//------- MODULES : setA
+#include "modules/codeLight.h"
+#include "modules/resetSwitch.h"
+#include "modules/lights.h"
+#include "modules/speakerAlarm.h"
+#include "modules/ultrasonic.h"
+//#include "modules/ambientlight.h"
+#include "modules/vibration.h"
+#include "modules/openReedSwitch.h"
+#include "modules/temperature.h"
 #include "modules/flame.h"
+#include "modules/laser.h"
+//DHT
+#include "libraries/dht.h"
+#include "libraries/dht.cpp"
+#include "modules/DHT11.h"
 
-////////////////////////////////////////////////////////
-void initModules(){
-  // initialize modules here
-  initFlame();
+
+
+//------- EXECUTE ALARM : conditions are set in each function
+void doAlarm(){
+    doLights();
+    doSpeaker();
+    //send SMS
 }
-void checkModules(){
-  // series of checks
+
+//------ WHAT TO CHECK : setB
+void checkForAlarm(){
+  checkUltraSonic();
+  checkTempSensor();
+  //checkAmbientLight();
+  checkReedSwitch();
+  checkVibration();
   checkFlame();
-}
-////////////////////////////////////////////////////////
-void setup() {
-  Serial.begin(9600);  
-  initAlarm();
-  initModules();
+  checkDHT11();
+  checkLaser();
+  checkResetButton();
 }
 
+//------ SET UP ROUTINES
+void setup() {
+  Serial.begin(9600);
+  initCodeLights();
+  initResetButton();
+  initLights();
+  initSpeaker();
+  initVibration();
+  initUltraSonic();
+  initTempSensor();
+  initReedSwitch();
+  initFlame();
+  initLaser();
+  initDHT11();
+}
+
+//------ KEEP CHECKING UNTIL ALARM
 void loop() {
-    delay(1000);
-    if(systemStatus == 15 && digitalRead(buttonPin) == HIGH){
-        // can only reset if had already been running in signalInterval iteration
-        Serial.println("button is High, resetting..");
-        systemStatus = 0;
-        setStatus(systemStatus);
-        setAlarmOff();
-        statusReceiver(0);
-        signalInterval = signalIntervalDefault;
-    }
-    
-    // SYSTEM STATUS TEST
-    /*systemStatus = systemStatus % 16;
-    if(systemStatus < 15){
-      setStatus(systemStatus);
-      systemStatus++;
-    }*/
-    
-    // main checker
-    checkModules();
-	  systemStatus = systemStatus < 1 ? 0 : systemStatus;
-    Serial.print("systemStatus: ");
-    Serial.println(systemStatus);
-    
-    setStatus(systemStatus);
-    
-    // counter
-    signalInterval--;
-    signalInterval = signalInterval < 0 ? 0 : signalInterval;
-    Serial.print("Interval: ");
-    Serial.println(signalInterval);
+  checkForAlarm();
+  if(systemUp == HIGH){
+    doAlarm();
+    doCodeLight();
+  }
 }
