+++ modules/flame.h
@@ -1,29 +1,27 @@
 
-//int flameDPin = 22; //digital
+int flameDPin = 22; //digital
 int flameAPin = 2; //analog
-int flameThreshold = 120;
+int flameThreshold = 100;
 
 void initFlame(){
-	//pinMode(flameDPin,INPUT);
+	pinMode(flameDPin,INPUT);
 }
 
 void checkFlame(){
-	//int flameDVal = digitalRead(flameDPin);
+	int flameDVal = digitalRead(flameDPin);
 	int flameAVal = analogRead(flameAPin);
 	
 	Serial.print("Flame = (A): ");
-	Serial.println(flameAVal);
-	//Serial.print(" (D): ");
-	//Serial.println(flameDVal);
+	Serial.print(flameAVal);
+	Serial.print(" (D): ");
+	Serial.println(flameDVal);
 	
-	//if(flameDVal == HIGH){
+	if(flameDVal == HIGH){
 		
-	//}
+	}
 	if(flameAVal < flameThreshold){
 		Serial.println("[[[[[[  FLAME ]]]]]");
-		systemStatus = 15;
-	}
-	else{
-		systemStatus = systemStatus > 12 ? systemStatus : 0;
+		codeShow(3);
+		setAlarm = HIGH;
 	}
 }
