--- modules/relaySwitch.h
@@ -1,12 +1,0 @@
-
-int relaySwitchPin = 6;
-
-void initRelaySwitch(){
-	pinMode(relaySwitchPin,OUTPUT);
-}
-void relaySwitchOn(){
-	digitalWrite(relaySwitchPin, HIGH);
-}
-void relaySwitchOff(){
-	digitalWrite(relaySwitchPin, LOW);
-}
